Separate full framework tests by resource/test type.
